% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biguessr.R
\name{biguessr}
\alias{biguessr}
\title{Automated Gender-Coding for Given Names}
\usage{
biguessr(name_vector)
}
\arguments{
\item{name_vector}{A character vector of given names to be processed.}
}
\value{
A tibble with columns
\itemize{
\item \code{given_name} – the original name in lower case
\item \code{gender}     – a factor with levels 1 (male), and 2 (female)
\item \code{method}     – “KNOWN” if the name was found in the predefined list, “GUESSED” if the coding was done by the LLM
}
}
\description{
This function uses a predefined list of given names and genders to infer gender from names. If any given name is not present in the predefined list, the large language model llama3.2 is used to guess the gender. This package is intended for use in the DACH region (Germany, Austria, Switzerland) and is based on historical baby names in those countries.
}
\examples{
\dontrun{
  names <- c("Jürgen","Franzi","Rudi","Pepe")
  biguessr(names)
}
}
