% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender_list}
\alias{gender_list}
\title{Full given names dataset}
\format{
A tibble with 24,414 rows and 2 columns:
\describe{
\item{given_name}{The given name}
\item{gender}{The gender associated with the name}
}
}
\source{
https://data.statistik.gv.at/web/meta.jsp?dataset=OGDEXT_VORNAMEN_1

https://www.babynames-stat.ch/de/
}
\usage{
gender_list
}
\description{
A dataset of given names and their associated gender. Compiled from official historic birth name data from Switzerland and Austria.
}
\keyword{datasets}
